refactored pr mutations.
