+++ src/resolvers/bounty/mutation.js
@@ -50,11 +50,8 @@ const Mutation = {
 	},
 	watchBounty: async (parent, args, { req, prisma }) => {
 		const signatureRegex = /signature=\w+/;
-		console.log(req.headers.cookie);
-		console.log(req.headers.cookie.match(signatureRegex));
 		const signature = req.headers.cookie.match(signatureRegex)[0].slice(10);
 		const address = ecdsaRecover(signature);
-		console.log(address, args.userAddress);
 		if (address !== args.userAddress) {
 			throw new AuthenticationError();
 		}
