+++ src/typeDefs/pr/mutation.js
@@ -5,7 +5,7 @@ const mutationDefs = gql`
 		createPr(thumbnail: String, prId: String!, bountyAddress: String!): PR!
 		updatePr(thumbnail: String, contributorIds: [String], prId: String!
 		): PR!
-		addContributor(prId: String, userId: String, address: String): PR
+		addContributor(prId: String, userId: String, address: String,  thumbnail:String): PR
 		removeContributor(prId: String, userId: String, address: String):PR
   }
 `;
